Create README.md
Applications of Principal Component Analysis (PCA)

A project work on the implementation of 2 applications of PCA

The first application illustrates how we can use PCA to project 2-D data onto its principal axes, meaning the two orthogonal directions that explain most of the variance in your data.

For the second application, we will use PCA to project higher dimensional data down to a lower dimensional feature space. This is an example of dimension reduction, a powerful technique that has multiple benefits, including reducing your model-building computational load and, in many cases, the accuracy of your model. PCA can help filter out redundant, linearly correlated variables and reduce the amount of noise in our data.

Using PCA to project 2-D data onto its principal axes

Here, we will illustrate how we can use PCA to transform your 2-D data to represent it in terms of its principal axes - the projection of our data onto the two orthogonal directions that explain most of the variance in the data.